requirements.txt
python-docx
openai
python-dotenv
selenium
smolagents
tiktoken
